Created AbstractEntity, Property, and Room Classes
