+++ src/main/java/com/example/burrowwebapp/models/AbstractEntity.java
@@ -1,0 +1,54 @@
+package com.example.burrowwebapp.models;
+
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+import javax.validation.constraints.NotBlank;
+import java.util.Objects;
+
+@MappedSuperclass
+public abstract class AbstractEntity
+{
+    @Id
+    @GeneratedValue
+    private int id;
+
+    @NotBlank
+    private String name;
+
+    public int getId(){
+        return id;
+    }
+
+    public String getName()
+    {
+        return name;
+    }
+
+    public void setName(String name)
+    {
+        this.name = name;
+    }
+
+    @Override
+    public String toString()
+    {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o)
+    {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        AbstractEntity that = (AbstractEntity) o;
+        return id == that.id;
+    }
+
+    @Override
+    public int hashCode()
+    {
+        return Objects.hash(id);
+    }
+}
