+++ src/main/java/com/example/burrowwebapp/models/Room.java
@@ -1,0 +1,11 @@
+package com.example.burrowwebapp.models;
+
+import javax.persistence.Entity;
+
+@Entity
+public class Room extends AbstractEntity
+{
+    public Room(){
+
+    }
+}
